requirements.txt
langchain
langchain-openai
pydantic
python-dotenv
requests
